src/app/+app/+clientes/create-client-modal/create-client-modal.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { CreateClientModalComponent } from './create-client-modal.component';

describe('Component: CreateClientModal', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [CreateClientModalComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([CreateClientModalComponent],
      (component: CreateClientModalComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(CreateClientModalComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(CreateClientModalComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-create-client-modal></app-create-client-modal>
  `,
  directives: [CreateClientModalComponent]
})
class CreateClientModalComponentTestController {
}

